refactor(.storybook): update root conifg decorator
