+++ .storybook/preview-head.html
@@ -1,9 +1,2 @@
-<style type="text/css">
-
-    :root {
-        --font-inter: 'Inter', sans-serif;
-    }
-
-</style>
 
 <link rel="preload" href="/fonts/inter/latin.woff2" as="font" type="font/woff2" crossorigin="anonymous" />
