+++ src/components/ThemeProvider.tsx
@@ -1,7 +1,17 @@
+import { useMemo } from "react"
+import merge from "lodash/merge"
 import { ThemeProvider as NextThemesProvider } from "next-themes"
 import type { ThemeProviderProps } from "next-themes/dist/types"
+import { ChakraBaseProvider, createLocalStorageManager } from "@chakra-ui/react"
+
+import customTheme from "@/@chakra-ui/theme"
 
 import { COLOR_MODE_STORAGE_KEY } from "@/lib/constants"
+
+import { useLocaleDirection } from "@/hooks/useLocaleDirection"
+import { mono } from "@/lib/fonts"
+
+const colorModeManager = createLocalStorageManager(COLOR_MODE_STORAGE_KEY)
 
 /**
  * Primary theming wrapper for use with color mode. Uses the theme provider
@@ -9,8 +19,13 @@ import { COLOR_MODE_STORAGE_KEY } from "@/lib/constants"
  *
  * Applied to _app.tsx as the main provider for the project, and supplied as the
  * primary decorator to Storybook.
+ *
+ * NOTE: This also includes the Chakra Provider. This will be removed after migration to ShadCN/Tailwind is complete
  */
 const ThemeProvider = ({ children }: Pick<ThemeProviderProps, "children">) => {
+  const direction = useLocaleDirection()
+
+  const theme = useMemo(() => merge(customTheme, { direction }), [direction])
   return (
     <NextThemesProvider
       attribute="data-theme"
@@ -19,7 +34,19 @@ const ThemeProvider = ({ children }: Pick<ThemeProviderProps, "children">) => {
       disableTransitionOnChange
       storageKey={COLOR_MODE_STORAGE_KEY}
     >
-      {children}
+      <ChakraBaseProvider theme={theme} colorModeManager={colorModeManager}>
+        {/* TODO: Can these CSS Vars be moved to `global.css`? */}
+        <style jsx global>
+          {`
+            :root {
+              --font-inter: Inter, sans-serif;
+              --font-mono: ${mono.style.fontFamily};
+            }
+          `}
+        </style>
+
+        {children}
+      </ChakraBaseProvider>
     </NextThemesProvider>
   )
 }
