--- tailwind/.storybook/manager-head.html
@@ -1,1 +1,0 @@
-<link rel="icon" type="image/png" href="/favicon.png" sizes="192x192" />
